Almost finished n minus one script

+++ n_minus_one/tests/test_evaluate_all.py
@@ -1,3 +1,4 @@
+import os
 from src import evaluate_all
 import shutil
 
@@ -5,7 +6,8 @@ import shutil
 def test_evaluate_all():
     """Test if the complete pipeline runs."""
     root = "tests/results"
-    shutil.rmtree(root)
+    if os.path.exists(root):
+        shutil.rmtree(root)
     evaluate_all(
         root,
         "tests/test_parameters.json",
@@ -13,6 +15,6 @@ def test_evaluate_all():
         graph_name="Macaque",
         has_weights=False,
         mlp_epochs=1,
-        embedder_epochs=1
+        embedder_epochs=10
     )
     shutil.rmtree(root)
